requirements.txt
numpy>=1.16.6
opencv-python==4.2.0.32
pyyaml>=5.3.1
rospkg>=1.2.3
roslibpy>=1.2.1
python-can>=3.3.4
